% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betarate_g.R
\name{betarate_g}
\alias{betarate_g}
\title{Sifted Data Utility Check}
\usage{
betarate_g(sdata, oridata, covtrue = c(10, 0.2, 5, 1, 2),
  form = as.formula("Y~x1+x2+x3+visit+(1|ID)"), family = gaussian(link
  = "identity"))
}
\arguments{
\item{sdata}{The Sifted data frame.}

\item{oridata}{The original data frame.}

\item{covtrue}{Ture or estimated coefficients of the target model.}

\item{form}{Model formula according to "covtrue" order.}

\item{family}{Family object to specify the link function for the outcome.}
}
\value{
\itemize{
   \item covrate - Absolute deviance between the sifted and original coefficients over the absolute value of original coefficients.
   \item covrate_t - Absolute deviance between the sifted and true coefficients over the absolute value of true coefficients.
   \item covrate_ori_t - Absolute deviance between the original and true coefficients over the absolute value of true coefficients.
   \item overlap - If the sifted and original CIs overlap.
   \item trueinci - If the true coefficients are in the sifted CIs.
   \item c1 - Confidence Intervals for the Sifted model.
   \item trueinorici - If the true coefficients are in the original CIs.
}
}
\description{
Check the deviance of sifted and original data parameter estimates under a parametric model. Model specification is needed for the check.
}
