% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calWeights.R
\name{calWeights}
\alias{calWeights}
\title{Calculate sampling weights for raw data imputation.}
\usage{
calWeights(
  data,
  lcolnames,
  ID,
  w = c("peri", "perij"),
  timevar,
  method = "param"
)
}
\arguments{
\item{data}{Original data (numerical variables only) with missing values.}

\item{lcolnames}{A vector of longitudinal variables names.}

\item{w}{Type of sampling weights or missingness level. "peri" is to consider weights on subject level, which means any subjects with partial missing would be excluded from complete cases. "perij" is to consider weights on subject and time level.
Only subjects with all time points missing would be excluded from complete cases.}

\item{timevar}{The time variable or cluster varaible name.}

\item{method}{Missingness model. If method = "param", the function utilize logistic regression ("peri") or GLMM ("perij") for missingness model.
If method = "nonparam", the function utilize random forest ("peri") for missingness model.}
}
\value{
A list contains two elements.
\itemize{
   \item Element 1 - Inversed probability of being observed.
   \item Element 2 - Row numbers of complete records in the original data.
}
#' @import dplyr glmmLasso glmnet doParallel lme4 missForest REEMtree MASS nlme
}
\description{
Use logistic regression or GLMM with LASSO regularity to model missingness for time-varying correlated data.
Output weights and full observation rows for each time-varying varaible.
}
