% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLink.R
\name{checkLink}
\alias{checkLink}
\title{Define link function for GLMM}
\usage{
checkLink(data, name)
}
\arguments{
\item{data}{A data frame in long format with outcome variable.}

\item{name}{the name of the outcome varaible.}
}
\value{
Proper link function.
}
\description{
Define link function for GLMM
}
