% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.dummy.R
\name{create.dummy}
\alias{create.dummy}
\title{Construct dummy variables for factor time-invariant vectors.}
\usage{
create.dummy(data, cs.names = NULL)
}
\arguments{
\item{data}{A data frame in long format with time-varying and time-invariant variables.}

\item{cs.names}{A vecotor of names for the time-invariant variables.}
}
\value{
A long format data with time-invariant factor/categorical variables changed to dummy variables.
}
\description{
Construct dummy variables for factor time-invariant vectors.
}
