% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVCF.R
\name{CVCF}
\alias{CVCF}
\title{Last value carry forward and next value carry backward combination for initiating imputation.}
\usage{
CVCF(data, lnames, timevar, ID)
}
\arguments{
\item{data}{A data frame in long format with missing value.}

\item{lnames}{A vector of the longitudinal varaible names.}

\item{timevar}{Name of the visit time variable.}

\item{ID}{Name of the ID variable.}
}
\description{
Last value carry forward and next value carry backward combination for initiating imputation.
}
