% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.imp.R
\name{diff.imp}
\alias{diff.imp}
\title{Calculating the difference between imputed and original values.}
\usage{
\method{diff}{imp}(ori, imp)
}
\arguments{
\item{ori}{Original numerical vector.}

\item{imp}{Imputed numerical vector.}
}
\value{
Sum of squared deviance between original and sifted vector over the sum of squared original vector.
}
\description{
Calculating the difference between imputed and original values.
}
