% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.lambda.R
\name{filter.lambda}
\alias{filter.lambda}
\title{Optimal lambda and corresponding final GLMM LASSO model}
\usage{
filter.lambda(data, outcome, ID, lambdaseq, range = 0,
  family = gaussian(link = "identity"), startPQL = TRUE)
}
\arguments{
\item{data}{A long format complete data frame.}

\item{outcome}{Outcome variable of interest.}

\item{ID}{Name of the ID variable in the dataset.}

\item{lambdaseq}{A vector of lambdas to consider.}

\item{range}{The next lambda sequence to consider, given the best lambda from current lambdaseq.}

\item{family}{Link function for the outcome.}

\item{startPQL}{Logical. If TRUE initiate parameter estimates are generated from glmmPQL. IF FALSE, initiate parameter estimates are 0s.}
}
\value{
\itemize{
   \item glm1 - The optimal GLMM under LASSO regularization with smallest BIC.
   \item seqs - The next Lambda sequence to be considered, when further search is needed.
}
}
\description{
Select the optimal lambda that minimizes BIC. The starting values of the parameters play an essential part in GLMM LASSO model fitting.
We first consider glmmPQL fitted values. Then, we consider 0s for sparse signal scenario.
}
