% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifv.R
\name{pifv}
\alias{pifv}
\title{Percent of Identical Feature Values}
\usage{
pifv(originaldata, sifteddata)
}
\arguments{
\item{originaldata}{Original data frame.}

\item{sifteddata}{Sifted data frame.}
}
\value{
A vector of Percent of Identical Feature Values for each row in the Sifted data frame.
}
\description{
Calculate the Percent of Identical Feature Values per record for the Sifted dataset.
}
