% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repSifter.R
\name{repSifter}
\alias{repSifter}
\title{DataSifter II Algorithm for Time-varying Data}
\usage{
repSifter(data, mispct, misw = "perij", lnames, timevar, ID,
  maxit = 10, crit = 0.05, cal.weights.method = "param")
}
\arguments{
\item{data}{A data frame contains original data to be processed. The data must be in long format. Missingness is allowed in time-varying varaibles.}

\item{mispct}{Percent of artificial missing that should be introduced for obfuscation. 20\%-30\% is recommended for utility preservation.}

\item{misw}{Type of sampling weights or missingness level. "peri" is to consider weights on subject level, which means any subjects with partial missing would be excluded from complete cases. "perij" is to consider weights on subject and time level. Only subjects with all time points missing would be excluded from complete cases.}

\item{lnames}{A vector of longitudinal variables names.}

\item{timevar}{The time variable or cluster varaible name.}

\item{ID}{Name of the ID variable in the dataset.}

\item{maxit}{Maximal iteration. The default is 10 times.}

\item{crit}{Critical value for the stopping criteria. The default is 0.05, which stops the algorithm when the absolute deviance of the imputed and original value is within 5\% of the original values.}

\item{cal.weights.method}{Raw data missingness model for calculating IPW weights. If method = "param", the function utilize logistic regression ("peri") or GLMM ("perij") for missingness model. If method = "nonparam", the function utilize random forest ("peri") for missingness model.}
}
\value{
\itemize{
  \item siftdata - Sifted data frame.
  \item data - Original data frame.
}
}
\description{
Create a informative privacy-preserving time-varying dataset that guarantees subjects' privacy while preserving the information contained in the original dataset.
}
