% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repSifterImp.R
\name{repSifterImp}
\alias{repSifterImp}
\title{Artificial Missing Introduce and Imputation Step}
\usage{
repSifterImp(data, mispct, lnames, timevar, ID, maxit = 10,
  crit = 0.05)
}
\arguments{
\item{data}{A data frame contains original data to be processed. The data must be in long format. Missingness is allowed in time-varying varaibles.}

\item{mispct}{Percent of artificial missing that should be introduced for obfuscation. 20\%-30\% is recommended for utility preservation.}

\item{lnames}{A vector of longitudinal variables names.}

\item{timevar}{The time variable or cluster varaible name.}

\item{ID}{Name of the ID variable in the dataset.}

\item{maxit}{Maximal iteration. The default is 10 times.}

\item{crit}{Critical value for the stopping criteria. The default is 0.05, which stops the algorithm when the absolute deviance of the imputed and original value is within 5\% of the original values.}

\item{misw}{Type of sampling weights or missingness level. "peri" is to consider weights on subject level, which means any subjects with partial missing would be excluded from complete cases. "perij" is to consider weights on subject and time level. Only subjects with all time points missing would be excluded from complete cases.}
}
\description{
Artificial Missing Introduce and Imputation Step
}
