% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{synthetic}
\alias{synthetic}
\title{Create fully synthetic longitudinal dataset.}
\usage{
synthetic(data, lnames, timevar, ID, N)
}
\arguments{
\item{data}{A data frame in long format.}

\item{lnames}{A vector of longitudinal variables names.}

\item{timevar}{The time variable or cluster varaible name.}

\item{ID}{Name of the ID variable in the dataset.}

\item{N}{The number of closest records to be considered as neighborhood.}
}
\value{
Fully synthetic Sifted data.
}
\description{
When data privacy is the major issue of time-varing data sharing, we consider this fully synthetic version of DataSifter II.
It obfuscate longitudinal data by generating random numbers within the neighborhood of a longitudinal variable at a given visit time.
Thus, the within-subject covariance structure might be changed by this process.
}
