% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_glmmlasso.R
\name{try_glmmlasso}
\alias{try_glmmlasso}
\title{Calculating the BIC of a GLMM lasso model, given specific lambda.}
\usage{
try_glmmlasso(lambdainp, data, family, form, Delta.start = NULL,
  q.start = NULL)
}
\arguments{
\item{lambdainp}{Lambda parameter for the GLMM lasso model.}

\item{data}{A data frame with complete longitudinal and cross-sectional data.}

\item{family}{A family object to specify the link function for outcome.}

\item{form}{Formula for GLMM lasso model.}

\item{Delta.start}{A scalar specifying starting value for GLMM lasso fixed effects parameters.}

\item{q.start}{A scalar or matrix of suitable dimension, specifying starting values for the random-effects variance-covariance matrix.}
}
\description{
Calculating the BIC of a GLMM lasso model, given specific lambda.
}
