% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlmmLasso.R
\name{wlmmLasso}
\alias{wlmmLasso}
\title{Weighted GlmmLASSO}
\usage{
wlmmLasso(x, y, yname, weights, rowid, ID = "ID", lvars,
  family = gaussian(link = "identity"), withlong = FALSE)
}
\arguments{
\item{x}{Data frame or matrix of candidate predictors time-invariant and longitudinal from the full data.}

\item{y}{Vector of longitudinal outcome of interest.}

\item{weights}{Sampling weights.}

\item{rowid}{Observed row id this will match with the length of weights.}

\item{lvars}{Longitudinal variable names.}

\item{family}{family object to specify the link function for outcome.}

\item{withlong}{a logical parameter to decide if the weighted linear mixed model depends on longitudinal predictors.}
}
\value{
The GLMM LASSO model under given weights.
}
\description{
GLMM LASSO with weights for each record. This can be used when we have a biased sample.
}
